+++ compiler-tests/src/test/data/diagnostic/bindingcontainer/TypeAliasesAreSpecifiedInDiagnostics.kt
@@ -1,0 +1,17 @@
+// RENDER_DIAGNOSTICS_FULL_TEXT
+// LANGUAGE: +NestedTypeAliases
+
+typealias UserId = String
+
+@DependencyGraph
+interface AppGraph {
+  typealias NestedUserId = String
+
+  val userId: UserId
+
+  @Binds fun String.<!PROVIDES_ERROR!>bind<!>(): UserId
+
+  @Binds fun String.<!PROVIDES_ERROR!>bindNested<!>(): NestedUserId
+
+  @Provides fun provideString(): String = "Hello"
+}